Initial commit
This is the initial commit of securitycenter-go,
after a 5-day initial sprint. Major changes from
this and securitycenter-cli is that this project
is being developed with a library driven development
mindset, creating avenues to access SecurityCenter
APIs natively in Go as opposed to creating a
command-line application first.

In order to meet the requirements of providing a
command-line interface to export and import mass
amounts of information from SecurityCenter, a CLI
companion application can be found in cmd/sc.